fix : RecentOrderConsumer 토픽명 에러 수정 및 PointService 재시도 로직 추가, 테스트 코드 작성
